LPTVA finished

+++ ObserverGraph.jl
@@ -11,7 +11,7 @@ struct ObsGraph
     infected_set::Set
 end
 
-function make_ER_obs_graph(N::Integer, p::Float16, frac_observers)::ObsGraph
+function make_ER_obs_graph(N::Integer, p::Float64, frac_observers)::ObsGraph
     @assert 0.0 <= p <= 1.0
     @assert 0.0 <= frac_observers <= 1.0
 
@@ -21,7 +21,6 @@ function make_ER_obs_graph(N::Integer, p::Float16, frac_observers)::ObsGraph
 end
 
 function make_BA_obs_graph(N::Int, n0::Int, k::Int, frac_observers)::ObsGraph
-    # @assert 0.0 <= p <= 1.0
     @assert 0.0 <= frac_observers <= 1.0
 
     g = Graphs.barabasi_albert(N, n0, k, complete=true)
